Moved more types into upvar.rs (now named closure.rs)
